+++ pluginmgr/__init__.py
@@ -3,10 +3,8 @@ from past.builtins import basestring
 from builtins import object
 
 import imp
-import collections
 import importlib
 import logging
-import munge.util
 import os
 import re
 import sys
@@ -106,27 +104,11 @@ class SearchPathImporter(object):
         else:
             mod = imp.load_source(name, filename)
 
-        self.log.debug("hook.loaded(%s) - %s", fullname, str(mod))
+        # don't need to check mod, both throw instead of returning None
 
-        if not mod:
-            raise ImportError(name)
+        self.log.debug("hook.loaded(%s) - %s", fullname, str(mod))
         sys.modules[fullname] = mod
         return mod
-
-
-# TODO move this to test / docs
-class PluginBase(object):
-    """
-    Example base class for plugins, set config and call init()
-    """
-    def __init__(self, config, *args, **kwargs):
-        self.config = config
-        self.args = args
-        self.kwargs = kwargs
-        self.init()
-
-    def init(self):
-        pass
 
 
 class PluginManager(object):
@@ -194,81 +176,3 @@ class PluginManager(object):
             self.log.debug("ImportError " + str(e))
 
         raise ValueError("unknown plugin '%s'" % typ)
-
-# config plugin only
-    def _ctor(self, typ, config, *args, **kwargs):
-        self.log.debug("ctor: self._instance %s", str(self._instance))
-        self.log.debug("ctor: self._class %s", str(self._class))
-        if typ in self._instance:
-            # get class type, copy config, override with passed config
-            obj = self._instance[typ]
-            cp = obj.config.copy()
-            munge.util.recursive_update(cp, config)
-            return type(obj)(cp, *args, **kwargs)
-        # try to load
-        return self.get_plugin_class(typ)(config, *args, **kwargs)
-        # FIXME - raise error, list configured class/instance
-
-# config plugin only
-    def new_plugin(self, config, *args, **kwargs):
-        """
-        instantiate a plugin
-        creates the object, stores it in _instance
-        """
-        typ = None
-        obj = None
-
-        # if type is defined, create a new instance
-        if 'type' in config:
-            typ = config['type']
-
-        # single key is overriding an existing plugin instance
-        elif isinstance(config, collections.Mapping) and len(config) == 1:
-            # get type name and shift out config to parent level
-            (typ, config) = list(config.items())[0]
-
-        obj = self._ctor(typ, config, *args, **kwargs)
-
-        # need to check for None, Greenlets return False
-        if obj is None:
-            raise ValueError("unable to instantiate plugin from %s" % str(config))
-
-        # store if named
-        if 'name' in config:
-            obj.name = config['name']
-            self._instance[obj.name] = obj
-        else:
-            # this could dupe on .name, make name=''?
-            obj.name = typ
-
-        return obj
-
-# config plugin only
-    def get_instance(self, node, *args, **kwargs):
-        """
-        get plugin instance from config node
-        *NOTE* returns an uninitialized instance if one isn't there
-        *NOTE* instantiated plugins without names remain anonymous
-               FIXME - why would instantiate() even process them
-        """
-        # string is a ref to an existing plugin instance
-        if isinstance(node, basestring):
-            if node in self._instance:
-                return self._instance[node]
-            # if not an instance, try for init with empty config
-            return self.new_plugin({'type': node}, *args, **kwargs)
-
-        if isinstance(node, collections.Mapping):
-            return self.new_plugin(node, *args, **kwargs)
-
-        raise ValueError("unable to parse plugin for output %s" % str(node))
-
-# config plugin only
-    def instantiate(self, config, *args, **kwargs):
-        """
-        takes plugin config (list under 'plugin') and instantiates defined
-        plugins
-        """
-        for plugin_config in config:
-            self.new_plugin(plugin_config, *args, **kwargs)
-