+++ tests/data/dynload/mod0.py
@@ -1,11 +1,7 @@
 
-import pluginmgr
 from pluginmgr_test import plugin
 
 
 @plugin.register('mod0')
-class Mod0(pluginmgr.PluginBase):
+class Mod0(object):
     pass
-
-
-