+++ tests/pluginmgr_test/__init__.py
@@ -1,9 +1,11 @@
 import os
 import pluginmgr
+import pluginmgr.config
 
 
 path = os.path.join(os.getcwd(), 'tests', 'data', 'dynload')
-plugin = pluginmgr.PluginManager('pluginmgr_test.plugins', path, False)
+#plugin = pluginmgr.PluginManager('pluginmgr_test.plugins', path, False)
+plugin = pluginmgr.config.ConfigPluginManager('pluginmgr_test.plugins', path, False)
 
 
 def set_create_loader(value=True):