+++ tests/pluginmgr_test/plugins/static0.py
@@ -1,9 +1,8 @@
 
-import pluginmgr
+from pluginmgr.config import PluginBase
 from pluginmgr_test import plugin
 
 
 @plugin.register('static0')
-class Static0(pluginmgr.PluginBase):
+class Static0(PluginBase):
     pass
-